backend/requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flower==2.0.1
frozenlist==1.5.0
humanize==4.12.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.0
magic-filter==1.0.12
multidict==6.2.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
